Refactor getEnvironment and gitInfo into more appropriate files
